refactor: added back the hazpy dependency and hazpy checkfor updates
I found that disabling the "manage.checkForHazPyUpdates()" in update.py does not work since in manage.py under "def startApp(self, app_path, update_path):" as the 'res' variable is 0 whether the virtual environment from the config.json exists or not, so it skips the "if res != 0" that would have installed hazpy.

And the manage.py appears to rely on the "def checkForHazPyUpdates(self):" to install hazpy whether the virtual env existed or not.

I'm trying it out on the fema machine with the hazpy dependency in the environment.yaml and the update.py retaining the checkforhazpy updates. I suppose FHIT could still have hazpy in its environment and it could be updated as hazpy is updated.